~Broke out functions into multiple files based functionality.
